requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.4
python-dotenv==1.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
langchain==0.2.6
langchain-openai==0.1.8
langchain-text-splitters==0.2.2
langchain_cohere
langchain_groq
